init(nodejs-basics): add function in the cjsToEsm.mjs, finish block modules
